Created data loader for page sidebar
