+++ docs/.vitepress/theme/data/config.ts
@@ -1,5 +1,6 @@
 import { SiteConfig } from "vitepress";
+import { ThemeConfig } from "../types";
 
-const config: SiteConfig = globalThis.VITEPRESS_CONFIG;
+const config: SiteConfig<ThemeConfig> = globalThis.VITEPRESS_CONFIG;
 
 export default config;
